+++ kernel/src/driver/pci/driver.rs
@@ -6,7 +6,7 @@ use crate::driver::base::device::{
     driver::{driver_manager, Driver},
 };
 
-use super::{dev_id::PciDeviceID, device::PciDevice, pci_bus};
+use super::{dev_id::PciDeviceID, device::PciDevice, subsys::pci_bus};
 
 /// # trait功能
 /// Pci驱动应该实现的trait
